pyproject.toml
[tool.poetry]
name = "d3r"
version = "0.1.0"
description = ".."
authors = ["xy"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
pandas = "^2.1.1"
wandb = "^0.15.12"
tqdm = "^4.66.1"
jupyter = "^1.0.0"
scikit-learn = "^1.3.2"
matplotlib = "^3.8.1"
kmeans-pytorch = "^0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"